+++ src/components/quantity-lessons-card/QuantityLessonChart/QuantityLessonsChart.constants.ts
@@ -1,21 +1,6 @@
-import React, { useMemo } from 'react'
-import { Bar } from 'react-chartjs-2'
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend
-} from 'chart.js'
-import { Box } from '@mui/material'
-import useBreakpoints from '~/hooks/use-breakpoints'
 import palette from '~/styles/app-theme/app.pallete'
 
-ChartJS.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend)
-
-const data = {
+export const data = {
   labels: [
     'JAN',
     'FEB',
@@ -42,7 +27,7 @@ const data = {
   ]
 }
 
-const options = {
+export const options = {
   responsive: true,
   maintainAspectRatio: false,
   plugins: {
@@ -75,22 +60,3 @@ const options = {
     }
   }
 }
-
-const QuantityLessonsChart = () => {
-  const { isMobile } = useBreakpoints()
-
-  const dataResponsive = useMemo(() => {
-    const obj = data
-    obj.datasets[0].barThickness = isMobile ? 20 : 28
-
-    return obj
-  }, [isMobile])
-
-  return (
-    <Box height={'181px'} key={+isMobile}>
-      <Bar data={dataResponsive} options={options} />
-    </Box>
-  )
-}
-
-export default QuantityLessonsChart
